Fixes related to obtain the avatar of a non contact

+++ app/src/main/java/mega/privacy/android/app/lollipop/listeners/ChatUserAvatarListener.java
@@ -8,6 +8,7 @@ import android.support.v7.widget.RecyclerView;
 
 import java.io.File;
 
+import mega.privacy.android.app.lollipop.megachat.chatAdapters.GroupCallAdapter;
 import mega.privacy.android.app.lollipop.megachat.chatAdapters.MegaListChatLollipopAdapter;
 import mega.privacy.android.app.utils.Util;
 import nz.mega.sdk.MegaApiJava;
@@ -38,38 +39,38 @@ public class ChatUserAvatarListener implements MegaRequestListenerInterface {
     @Override
     public void onRequestFinish(MegaApiJava api, MegaRequest request, MegaError e) {
         log("onRequestFinish(): "+e.getErrorCode());
-        if (e.getErrorCode() == MegaError.API_OK){
-            boolean avatarExists = false;
+        if(request.getType() == MegaRequest.TYPE_GET_ATTR_USER) {
 
-            if(holder instanceof MegaListChatLollipopAdapter.ViewHolderChatList){
-                if(((MegaListChatLollipopAdapter.ViewHolderNormalChatList)holder)!=null && ((MegaListChatLollipopAdapter.ViewHolderNormalChatList)holder).getContactMail()!=null && request.getEmail()!=null){
-                    if (((MegaListChatLollipopAdapter.ViewHolderNormalChatList)holder).getContactMail().compareTo(request.getEmail()) == 0){
-                        File avatar = null;
-                        if (context.getExternalCacheDir() != null){
-                            avatar = new File(context.getExternalCacheDir().getAbsolutePath(), ((MegaListChatLollipopAdapter.ViewHolderNormalChatList)holder).getContactMail() + ".jpg");
-                        }
-                        else{
-                            avatar = new File(context.getCacheDir().getAbsolutePath(), ((MegaListChatLollipopAdapter.ViewHolderNormalChatList)holder).getContactMail() + ".jpg");
-                        }
-                        Bitmap bitmap = null;
-                        if (avatar.exists()){
-                            if (avatar.length() > 0){
-                                BitmapFactory.Options bOpts = new BitmapFactory.Options();
-                                bOpts.inPurgeable = true;
-                                bOpts.inInputShareable = true;
-                                bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-                                if (bitmap == null) {
-                                    avatar.delete();
-                                }
-                                else{
-                                    ((MegaListChatLollipopAdapter.ViewHolderNormalChatList)holder).setImageView(bitmap);
+            if (e.getErrorCode() == MegaError.API_OK) {
+                boolean avatarExists = false;
+
+                if (holder instanceof MegaListChatLollipopAdapter.ViewHolderChatList) {
+                    if (((MegaListChatLollipopAdapter.ViewHolderNormalChatList) holder) != null && ((MegaListChatLollipopAdapter.ViewHolderNormalChatList) holder).getContactMail() != null && request.getEmail() != null) {
+                        if (((MegaListChatLollipopAdapter.ViewHolderNormalChatList) holder).getContactMail().compareTo(request.getEmail()) == 0) {
+                            File avatar = null;
+                            if (context.getExternalCacheDir() != null) {
+                                avatar = new File(context.getExternalCacheDir().getAbsolutePath(), ((MegaListChatLollipopAdapter.ViewHolderNormalChatList) holder).getContactMail() + ".jpg");
+                            } else {
+                                avatar = new File(context.getCacheDir().getAbsolutePath(), ((MegaListChatLollipopAdapter.ViewHolderNormalChatList) holder).getContactMail() + ".jpg");
+                            }
+                            Bitmap bitmap = null;
+                            if (avatar.exists()) {
+                                if (avatar.length() > 0) {
+                                    BitmapFactory.Options bOpts = new BitmapFactory.Options();
+                                    bOpts.inPurgeable = true;
+                                    bOpts.inInputShareable = true;
+                                    bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
+                                    if (bitmap == null) {
+                                        avatar.delete();
+                                    } else {
+                                        ((MegaListChatLollipopAdapter.ViewHolderNormalChatList) holder).setImageView(bitmap);
+                                    }
                                 }
                             }
                         }
+                    } else {
+                        log("Adapter cannot be updated - null");
                     }
-                }
-                else{
-                    log("Adapter cannot be updated - null");
                 }
             }
         }
